+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 
 # dictionary sorter change log
 
+#### 0.7.1
+- backwards compatability: strictness converted to leniency
+
 ### 0.7.0
-- you can now input repeatable letters just by putting more than 1 of that letter in the `letters` array, and it will only repeat the amount of times you put in that letter.
+- you can now input repeatable letters just by putting more than 1 of that letter in the `letters` array, and it will only repeat the amount of times you put in that letter. (TempLetters)
+- fixed bug: duplicate outputs. 
 
 #### UI-text 0.1 - 0.6.5
 - this is a text based user interface to make a settings file for dictionary sorter version 0.6.5. It is probably still better to directly edit settings.pyw though.